add useExtendTree hook
