+++ packages/binding/src/accessorPropagation/index.ts
@@ -20,6 +20,7 @@ export * from './useEntityListSubTreeParameters'
 export * from './useEntitySubTree'
 export * from './useEntitySubTreeParameters'
 export * from './useEnvironment'
+export * from './useExtendTree'
 export * from './useField'
 export * from './useGetEntityByKey'
 export * from './useGetEntityListSubTree'
